requirements.txt
transformers>=4.45.0
torch>=2.7.0
torchvision>=0.22.0
pillow>=11.0.0
numpy>=2.2.0
